# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My OpenApi Documentation",
        "description": "My OpenApi description",
        "contact": {
            "email": "markusmak@gmail.com"
        },
        "license": {
            "name": "Proprietary",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "My API Server"
        }
    ],
    "paths": {
        "/api/cadastre": {
            "get": {
                "tags": [
                    "Cadastres"
                ],
                "summary": "Get list of cadastres",
                "description": "Returns list of cadastres",
                "operationId": "getCadastres",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cadastres"
                ],
                "summary": "Create a new cadastre",
                "description": "Create a new cadastre entry in the database by providing the required information.",
                "operationId": "f4dc798006d98204e87ad75c81627217",
                "requestBody": {
                    "description": "Cadastre data to be stored",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the newly created cadastre data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastreRequest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Validation error message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cadastre/{id}": {
            "get": {
                "tags": [
                    "Cadastres"
                ],
                "summary": "Get a cadastre by ID",
                "description": "Retrieve the cadastre details by providing its ID.",
                "operationId": "352e9ff2c2a8d955e86a9823325f140b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cadastre to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the cadastre data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastreRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cadastre not found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Cadastre not found"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cadastres"
                ],
                "summary": "Update a cadastre by ID",
                "description": "Update the information of a cadastre by providing its ID.",
                "operationId": "d001717f88199d9daec35dae78e35be6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cadastre to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated cadastre data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the updated cadastre data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastreRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cadastre not found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Cadastre not found"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cadastres"
                ],
                "summary": "Delete a cadastre by ID",
                "description": "Delete a cadastre from the database by providing its ID.",
                "operationId": "ee147b43cd688671ec68a49bc764e437",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cadastre to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation. No content."
                    },
                    "404": {
                        "description": "Cadastre not found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Cadastre not found"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CadastreRequest": {
                "title": "CadastreRequest",
                "description": "Request schema for Cadastre",
                "required": [
                    "name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "description": "Phone number of the user",
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
...
        }
    }
}